fix(tyr): address review findings for NIU-614 ravn outcome ingestion
- Add max_retries guard in handle_ravn_outcome retry verdict path;
  when retries exhausted transitions to FAILED matching _handle_ci_failure
  pattern (Finding 1)
- Add owner_id field to RavnOutcomeConfig (default "api") and use
  settings.ravn_outcome.owner_id in main.py, removing coupling to
  event_triggers config (Finding 2)
- Move StubTracker, StubTrackerFactory, and make_raid into shared
  tests/test_tyr/stubs.py; import from there in ravn outcome test file
  (Finding 3)
- Add test for retry-exhausted-to-failed path
