+++ tests/test_tyr/stubs.py
@@ -5,10 +5,23 @@ from __future__ import annotations
 from collections.abc import AsyncGenerator
 from datetime import UTC, datetime
 from typing import Any
-from uuid import UUID
+from uuid import UUID, uuid4
 
-from tyr.domain.models import Phase, Raid, RaidStatus, Saga, SagaStatus
+from tyr.domain.models import (
+    ConfidenceEvent,
+    Phase,
+    PhaseStatus,
+    Raid,
+    RaidStatus,
+    Saga,
+    SagaStatus,
+    SessionMessage,
+    TrackerIssue,
+    TrackerMilestone,
+    TrackerProject,
+)
 from tyr.ports.saga_repository import SagaRepository
+from tyr.ports.tracker import TrackerPort
 from tyr.ports.volundr import (
     ActivityEvent,
     PRStatus,
@@ -193,3 +206,226 @@ class StubVolundrFactory:
 
     async def primary_for_owner(self, owner_id: str) -> VolundrPort | None:
         return self._volundr
+
+
+# ---------------------------------------------------------------------------
+# Tracker stubs (used by ravn outcome tests and review engine tests)
+# ---------------------------------------------------------------------------
+
+_NOW = datetime.now(UTC)
+_DEFAULT_OWNER = "test-owner"
+_DEFAULT_TRACKER_ID = "raid-tracker-001"
+_DEFAULT_SESSION = "sess-ravn-001"
+
+
+class StubTracker(TrackerPort):
+    """Minimal in-memory tracker for Tyr unit tests."""
+
+    def __init__(self, raid: Raid | None = None) -> None:
+        self._raid = raid
+        self._raids_by_session: dict[str, Raid] = {}
+        if raid is not None:
+            self._raids_by_id = {raid.tracker_id: raid}
+            if raid.session_id:
+                self._raids_by_session[raid.session_id] = raid
+        else:
+            self._raids_by_id: dict[str, Raid] = {}
+        self.confidence_events: dict[str, list[ConfidenceEvent]] = {}
+        self.phase: Phase | None = None
+        self.saga: Saga | None = None
+        self._phases: list[Phase] = []
+        self._all_merged: bool = False
+        self.closed_raids: list[str] = []
+
+    # -- CRUD: create entities --
+
+    async def create_saga(self, saga: Saga, *, description: str = "") -> str:
+        return saga.tracker_id
+
+    async def create_phase(self, phase: Phase, *, project_id: str = "") -> str:
+        return phase.tracker_id
+
+    async def create_raid(self, raid: Raid, *, project_id: str = "", milestone_id: str = "") -> str:
+        self._raids_by_id[raid.tracker_id] = raid
+        if raid.session_id:
+            self._raids_by_session[raid.session_id] = raid
+        return raid.tracker_id
+
+    async def update_raid_state(self, raid_id: str, state: RaidStatus) -> None:
+        pass
+
+    async def close_raid(self, raid_id: str) -> None:
+        self.closed_raids.append(raid_id)
+
+    # -- Read --
+
+    async def get_saga(self, saga_id: str) -> Saga:
+        if self.saga is None:
+            raise ValueError("No saga")
+        return self.saga
+
+    async def get_phase(self, tracker_id: str) -> Phase:
+        if self.phase is None:
+            raise ValueError("No phase")
+        return self.phase
+
+    async def get_raid(self, tracker_id: str) -> Raid:
+        raid = self._raids_by_id.get(tracker_id)
+        if raid is None:
+            raise ValueError(f"Raid not found: {tracker_id}")
+        return raid
+
+    async def list_pending_raids(self, phase_id: str) -> list[Raid]:
+        return []
+
+    async def list_projects(self) -> list[TrackerProject]:
+        return []
+
+    async def get_project(self, project_id: str) -> TrackerProject:
+        raise NotImplementedError
+
+    async def list_milestones(self, project_id: str) -> list[TrackerMilestone]:
+        return []
+
+    async def list_issues(
+        self, project_id: str, milestone_id: str | None = None
+    ) -> list[TrackerIssue]:
+        return []
+
+    # -- Raid progress --
+
+    async def update_raid_progress(
+        self,
+        tracker_id: str,
+        *,
+        status: RaidStatus | None = None,
+        session_id: str | None = None,
+        confidence: float | None = None,
+        pr_url: str | None = None,
+        pr_id: str | None = None,
+        retry_count: int | None = None,
+        reason: str | None = None,
+        owner_id: str | None = None,
+        phase_tracker_id: str | None = None,
+        saga_tracker_id: str | None = None,
+        chronicle_summary: str | None = None,
+        reviewer_session_id: str | None = None,
+        review_round: int | None = None,
+    ) -> Raid:
+        raid = self._raids_by_id.get(tracker_id)
+        if raid is None:
+            raise ValueError(f"Raid not found: {tracker_id}")
+        events = self.confidence_events.get(tracker_id, [])
+        new_confidence = events[-1].score_after if events else raid.confidence
+        updated = Raid(
+            id=raid.id,
+            phase_id=raid.phase_id,
+            tracker_id=raid.tracker_id,
+            name=raid.name,
+            description=raid.description,
+            acceptance_criteria=raid.acceptance_criteria,
+            declared_files=raid.declared_files,
+            estimate_hours=raid.estimate_hours,
+            status=status if status is not None else raid.status,
+            confidence=confidence if confidence is not None else new_confidence,
+            session_id=session_id if session_id is not None else raid.session_id,
+            branch=raid.branch,
+            chronicle_summary=raid.chronicle_summary,
+            pr_url=pr_url if pr_url is not None else raid.pr_url,
+            pr_id=pr_id if pr_id is not None else raid.pr_id,
+            retry_count=retry_count if retry_count is not None else raid.retry_count,
+            created_at=raid.created_at,
+            updated_at=datetime.now(UTC),
+            reviewer_session_id=(
+                reviewer_session_id if reviewer_session_id is not None else raid.reviewer_session_id
+            ),
+            review_round=review_round if review_round is not None else raid.review_round,
+        )
+        self._raids_by_id[tracker_id] = updated
+        if updated.session_id:
+            self._raids_by_session[updated.session_id] = updated
+        return updated
+
+    async def get_raid_progress_for_saga(self, saga_tracker_id: str) -> list[Raid]:
+        return list(self._raids_by_id.values())
+
+    async def get_raid_by_session(self, session_id: str) -> Raid | None:
+        return self._raids_by_session.get(session_id)
+
+    async def list_raids_by_status(self, status: RaidStatus) -> list[Raid]:
+        return [r for r in self._raids_by_id.values() if r.status == status]
+
+    async def get_raid_by_id(self, raid_id: UUID) -> Raid | None:
+        return next((r for r in self._raids_by_id.values() if r.id == raid_id), None)
+
+    async def add_confidence_event(self, tracker_id: str, event: ConfidenceEvent) -> None:
+        self.confidence_events.setdefault(tracker_id, []).append(event)
+
+    async def get_confidence_events(self, tracker_id: str) -> list[ConfidenceEvent]:
+        return self.confidence_events.get(tracker_id, [])
+
+    async def all_raids_merged(self, phase_tracker_id: str) -> bool:
+        return self._all_merged
+
+    async def list_phases_for_saga(self, saga_tracker_id: str) -> list[Phase]:
+        return self._phases
+
+    async def update_phase_status(self, phase_tracker_id: str, status: PhaseStatus) -> Phase | None:
+        return None
+
+    async def get_saga_for_raid(self, tracker_id: str) -> Saga | None:
+        return self.saga
+
+    async def get_phase_for_raid(self, tracker_id: str) -> Phase | None:
+        return self.phase
+
+    async def get_owner_for_raid(self, tracker_id: str) -> str | None:
+        return _DEFAULT_OWNER
+
+    async def save_session_message(self, message: SessionMessage) -> None:
+        pass
+
+    async def get_session_messages(self, tracker_id: str) -> list[SessionMessage]:
+        return []
+
+    async def attach_issue_document(self, issue_id: str, title: str, content: str) -> str:
+        return "doc-1"
+
+
+class StubTrackerFactory:
+    def __init__(self, tracker: StubTracker) -> None:
+        self._tracker = tracker
+
+    async def for_owner(self, owner_id: str) -> list[StubTracker]:
+        return [self._tracker]
+
+
+def make_raid(
+    *,
+    status: RaidStatus = RaidStatus.REVIEW,
+    confidence: float = 0.5,
+    session_id: str | None = _DEFAULT_SESSION,
+    retry_count: int = 0,
+    tracker_id: str = _DEFAULT_TRACKER_ID,
+) -> Raid:
+    """Build a minimal Raid for use in tests."""
+    return Raid(
+        id=uuid4(),
+        phase_id=uuid4(),
+        tracker_id=tracker_id,
+        name="test-raid",
+        description="A test raid",
+        acceptance_criteria=[],
+        declared_files=[],
+        estimate_hours=1.0,
+        status=status,
+        confidence=confidence,
+        session_id=session_id,
+        branch=None,
+        chronicle_summary=None,
+        pr_url=None,
+        pr_id=None,
+        retry_count=retry_count,
+        created_at=_NOW,
+        updated_at=_NOW,
+    )
